Bug fix
If there are many GFF under a same directory and we run `python -m jcvi.annotation.stats genestats`  for each of them, we will get unexpected result because the origin code will generate one "transcript.sizes" file. It is better to generate "transcript.sizes" with the filename of GFF as prefix.